Move sayto below say
If, for example, you type 'say hello' and send it to the mud, you will
get "say what and to whom" (or somesuch) back. This is because the command
'sayto' is higher on the priority list, than 'say'. (This is also why in
some MOBprogs, you aren't seeing things that mobs are supposed to be saying.)